tests: lib: hw_unique_key_tfm: Fix test
The key derivation used in TF-M is changed and
this change updates the tests to follow the new
key derivation method. It was also noticed
that the current nrfjprog doesn't allow to
write to the KMU region using the program
argument so this was updated to use the memory
write argument.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>